extract createScaleColors
