+++ app/javascript/gobierto_visualizations/webapp/lib/utils.js
@@ -1,5 +1,8 @@
 import { csv } from "d3-fetch";
+import { scaleOrdinal } from 'd3-scale';
 import axios from "axios";
+
+const d3 = { scaleOrdinal }
 
 const endPointGobiertoData = `/api/v1/data/data.json`
 
@@ -56,4 +59,14 @@ export function normalizeString(string) {
     .replace(/[.,()\s]/g, '')
     .toLowerCase();
   return slug
+}
+
+export function createScaleColors(values, arrayDomain) {
+  let colorsGobiertoExtend = ["#12365b", "#118e9c", "#ff766c", "#f7b200", "#158a2c", "#94d2cf", "#3a78c3", "#15dec5", "#6a7f2f", "#55f17b"]
+  colorsGobiertoExtend = colorsGobiertoExtend.slice(0, values)
+
+  const colors = d3.scaleOrdinal()
+    .domain(arrayDomain)
+    .range(colorsGobiertoExtend);
+  return colors;
 }
